Bugs, resources
Fixed grass walking bug (thanks lochie)
added natural resource spawning
fixed building system
need to fix allowed & blocked colours on building

+++ Brods Survival 3/Assets/Scripts/Gathering/GatherableObject.cs
@@ -7,7 +7,7 @@ public class GatherableObject : MonoBehaviour
     public GatherDataSO[] gatherDatas;
     public int hits;
     public ItemSO[] prefferedTools;
-    public int toolMultiplier = 2;
+    public float toolMultiplier = 2;
